Refactored canvas.js, added an example/test file
Made Rectangle, Image, Text and NumText seperate from the canvas object.

Made all objects local, except canvas.

Made all objects attributes of canvas.

Turned the resource loading tools into a constructor.

Made mouse input into a constructor, made the Canvas constructor always create
an instance of it.

Before, "input.keys" would only be set by supporting browsers. "canvas.keys",
which supercedes it, will now be set using the value of keyCode, and the object
keyMap.

Added a DT object, for calculating differences in time between events, and
calculating FPSes. Canvas has an attribute "dt", which is an instance of it

Removed loop, and all associated objects. Figured out how to do without them.

Commented out the code associated with touch input. Will hopefully reimplement
it, except as part of the mouse object, more inline with the coding style of the
rest of the library.

It was always just a copy paste in from a website. The old
one only ever did a bad job of acting as a mouse. Any reimplentation should
take better advantage of touching, whilst allowing simple usage on the part of
developers.